scripts/requirements.txt
requests==2.31.0
supabase==2.3.0
anthropic==0.20.0
yfinance==0.2.36
feedparser==6.0.10
python-dotenv==1.0.0
pytz==2024.1
